DHFPROD-4519: Add Model endpoint
